DemoSchoolBot/requirements.txt
aiogram==3.4.1
fastapi==0.111.0
uvicorn==0.30.0
asyncpg==0.29.0
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
redis==5.0.4
python-dotenv==1.0.1
pytest==8.2.0
Jinja2==3.1.3
starlette==0.37.2
qrcode==7.4.2
pydub==0.25.1
reportlab==4.1.0
openai==1.30.1
